Time: 3 ms (93.22%) | Memory: 8.5 MB (36.13%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-    long rev_num = 0; 
-    long original =x;
+    long long rev_num = 0; 
+    long long original =x;
     while (x > 0) { 
         rev_num = rev_num * 10 + x % 10; 
         x = x / 10;
